e2e canRunMigrations flag

+++ src/nestjs/src/@share/testing/helpers.ts
@@ -1,17 +1,37 @@
 import { INestApplication } from "@nestjs/common";
+import { ConfigService } from "@nestjs/config";
 import { getConnectionToken } from "@nestjs/sequelize";
 import { Test, TestingModule } from "@nestjs/testing";
+import { migrator } from "../../../../@core/src/@seedwork/infra/db/sequelize";
 import { AppModule } from "../../app.module";
 import { applyGlobalConfig } from "../../global-config";
 
 export function startApp({ beforeInit }: { beforeInit?: (app: INestApplication) => void } = {}) {
   let _app: INestApplication;
+  let canRunMigrations: boolean;
   beforeEach(async () => {
     const moduleBuilder: TestingModule = await Test.createTestingModule({
       imports: [AppModule],
     }).compile();
+
+    canRunMigrations = !moduleBuilder.get(ConfigService)
+      .get("DB_AUTO_LOAD_MODELS");
+
     const sequelize = moduleBuilder.get(getConnectionToken());
-    await sequelize.sync({ force: true });
+
+    try {
+      if (canRunMigrations) {
+        const umzug = migrator(sequelize, { logger: undefined });
+        await umzug.down({ to: 0 as any });
+        await umzug.up();
+      } else {
+        await sequelize.sync({ force: true });
+      }
+    } catch (e) {
+      console.error(e);
+      throw e;
+    }
+
     _app = moduleBuilder.createNestApplication();
     applyGlobalConfig(_app);
     beforeInit && beforeInit(_app);
